feat: add get rahkaran balance toghether with bank

+++ app/Services/Banking/BankAdapterInterface.php
@@ -6,10 +6,10 @@ interface BankAdapterInterface
 {
     public function setAccount(array $credentials): self;
 
-    public function getBalance(): float;
+    public function getBalance(): int;
 
     /**
-     * @return array{accountNumber: string, balance: float, todayDepositAmount: float, todayWithdrawAmount: float, currency: string}
+     * @return array{accountNumber: string, balance: int, todayDepositAmount: int, todayWithdrawAmount: int, currency: string}
      */
     public function getAccountBalance(): array;
 }
